LayoutTests/fast/viewport/viewport-77.html

<head>
    <title>"yes;" means "no".</title>
    <meta name="viewport" content="width=10 ;,initial-scale=yes;">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
